커스텀 마스킹 기능 및 헤더 필터링 기능 필요 : feat : 마스킹 기능 및 커스텀 헤더 필터링 기능 어노테이션에서 별도로 설정 오버라이드기능 추가 및 suhsaechan.kr 로 도메인명 수정 https://github.com/Cassiiopeia/suh-logger/issues/50
