+++ src/main/java/kr/suhsaechan/suhlogger/annotation/LogTime.java
@@ -1,0 +1,11 @@
+package kr.suhsaechan.suhlogger.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(value = RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface LogTime {
+}
