--- src/main/java/me/suhsaechan/suhlogger/SuhLoggerApplication.java
@@ -1,15 +1,0 @@
-package me.suhsaechan.suhlogger;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-
-@SpringBootApplication
-@EnableAspectJAutoProxy
-public class SuhLoggerApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(SuhLoggerApplication.class, args);
-  }
-
-}
