+++ src/main/resources/META-INF/spring.factories
@@ -1,7 +1,7 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  me.suhsaechan.suhlogger.config.SuhLoggerAutoConfiguration
+  kr.suhsaechan.suhlogger.config.SuhLoggerAutoConfiguration
 
 # Disable Spring Boot Logging
 org.springframework.boot.logging.LoggingSystemFactory=\
-  me.suhsaechan.suhlogger.config.NoOpLoggingSystemFactory
+  kr.suhsaechan.suhlogger.config.NoOpLoggingSystemFactory
