+++ src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-me.suhsaechan.suhlogger.config.SuhLoggerAutoConfiguration
+kr.suhsaechan.suhlogger.config.SuhLoggerAutoConfiguration
