+++ src/test/java/kr/suhsaechan/suhlogger/SuhLoggerApplicationTests.java
@@ -1,0 +1,15 @@
+package kr.suhsaechan.suhlogger;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
+
+@SpringBootTest
+@EnableAspectJAutoProxy
+class SuhLoggerApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
