--- src/test/java/me/suhsaechan/suhlogger/SuhLoggerApplicationTests.java
@@ -1,15 +1,0 @@
-package me.suhsaechan.suhlogger;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-
-@SpringBootTest
-@EnableAspectJAutoProxy
-class SuhLoggerApplicationTests {
-
-  @Test
-  void contextLoads() {
-  }
-
-}
