+++ version.yml
@@ -34,7 +34,7 @@
 # - 버전은 항상 높은 버전으로 자동 동기화됩니다
 # ===================================================================
 
-version: "1.4.2"
+version: "2.0.0"
 version_code: 2 # app build number
 project_type: "spring" # spring, flutter, react, react-native, react-native-expo, node, python, basic
 metadata:
